refactored code, added some init comments
