+++ src/AppComponents/DatabasePage/AudioDescriptionPage/AudioDescripitonPage.js
@@ -3,19 +3,20 @@ import ReactSelect from "react-select";
 import ReactSwitch from "react-switch";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-import './MLDescriptionPage.css'
+import './AudioDescripitonPage.css'
 
-import * as states from './States.js';
+import * as states from '../../AppTabStates.js';
 
-class MLDescriptionPage extends React.Component {
+class AudioDescripitonPage extends React.Component {
+
     render() {
         let audioName = `${this.props.audio == null ? 'none' : this.props.audio.name}`;
         return(
             <div>
-                MLDescription {audioName}
+                Audio Description {audioName}
             </div>
         );
     }
 }
 
-export default MLDescriptionPage
+export default AudioDescripitonPage
