+++ src/AppComponents/DatabasePage/MLDescriptionPage/MLDescriptionPage.js
@@ -3,20 +3,19 @@ import ReactSelect from "react-select";
 import ReactSwitch from "react-switch";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
-import './AudioDescripitonPage.css'
+import './MLDescriptionPage.css'
 
-import * as states from './States.js';
+import * as states from '../../AppTabStates.js';
 
-class AudioDescripitonPage extends React.Component {
-
+class MLDescriptionPage extends React.Component {
     render() {
         let audioName = `${this.props.audio == null ? 'none' : this.props.audio.name}`;
         return(
             <div>
-                Audio Description {audioName}
+                MLDescription {audioName}
             </div>
         );
     }
 }
 
-export default AudioDescripitonPage
+export default MLDescriptionPage
