+++ src/AppComponents/HardwarePage/HardwarePage.css
